% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoF.f.graph.R
\name{funct_graph}
\alias{funct_graph}
\title{Create frequency spectra}
\usage{
funct_graph(data, vcf, outfile)
}
\arguments{
\item{data}{df generated from either lof_funct.R or lof_pos.R}

\item{vcf}{vcf file from which lof_funct.R was ran on}

\item{outfile}{file path for the output of the graph}
}
\value{
null
}
\description{
This function generates a site-frequnecy spectra based on a dataframe of data created
from the lof_funct function.
}
