% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_vcf.R
\name{load_vcf}
\alias{load_vcf}
\title{Load Vcf
This functions loads a vcf file and formats it to be readable in R.}
\usage{
load_vcf(infile)
}
\arguments{
\item{infile}{path to infile}
}
\value{
formatted vcf file
}
\description{
Load Vcf
This functions loads a vcf file and formats it to be readable in R.
}
