% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lof_funct.R
\name{lof_funct}
\alias{lof_funct}
\title{Count loss of function mutations based on functional status.}
\usage{
lof_funct(x, y)
}
\arguments{
\item{x}{formatted vcf file}

\item{y}{File path for .csv file}
}
\value{
data frame with gene name and the LoF mutations within the gene
}
\description{
This function counts LoF mutations based on functional status and
outputs a df and csv file with the data. This assumes the function load_vcf was run on
a vcf file and if you use the package to generate a SFS, you assign the output of the function to a variable
in your environment
}
