% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lof_genes_funct.R
\name{lof_genes_funct}
\alias{lof_genes_funct}
\title{Count loss of function mutations based on functional status based on a gene list provided by the user.}
\usage{
lof_genes_funct(x, y, gt)
}
\arguments{
\item{x}{formatted vcf file}

\item{y}{File path for .csv file}

\item{gt}{txt file containing list of genes of interest}
}
\value{
data frame with gene name and the LoF mutations within the gene
}
\description{
This function counts LoF mutations based on functional status and
outputs a df and csv file with the data. This assumes the function load_vcf was run on
a vcf file and if you use the package to generate a SFS, you should assign the output of the function to a variable
in your environment
}
