% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lof_select_funct.R
\name{lof_select_funct}
\alias{lof_select_funct}
\title{Count loss of function mutations based on functional status and which specific mutations the user selects.}
\usage{
lof_select_funct(x, y, mut1, mut2, mut3 = "N/A,DNE", mut4 = "N/A,DNA", z)
}
\arguments{
\item{x}{formatted vcf file}

\item{y}{File path for .csv file}

\item{mut1}{Mutational type as seen in vcf}

\item{mut2}{Optional additional mutational type}

\item{mut3}{Optional additional mutational type}

\item{mut4}{Optional additional mutational type}

\item{z}{optional file path for potential mutations that may not be LoF and need further investigation}
}
\value{
data frame with gene name and the LoF mutations within the gene
}
\description{
This function counts LoF mutations based on functional status except only counting mutations
that the user specifices and outputs a df and csv file with the data. This assumes the function load_vcf was run on
a vcf file and if you use the package to generate a SFS, you assign the output of the function to a variable
in your environment
}
